+++ app/models/sheet_unlock_request.rb
@@ -27,6 +27,7 @@ class SheetUnlockRequest < ApplicationRecord
   def send_unlock_request_emails
     return unless EMAILS_ENABLED
     project_editors.each do |editor|
+      next unless sheet.project.emails_enabled?(editor)
       UserMailer.sheet_unlock_request(self, editor).deliver_now
     end
   end