+++ db/migrate/20161006201251_remove_email_notifications_from_users.rb
@@ -1,0 +1,5 @@
+class RemoveEmailNotificationsFromUsers < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :users, :email_notifications, :text
+  end
+end